+++ src/main/java/io/github/lokka30/levelledmobs/MobDataManager.java
@@ -1,5 +1,6 @@
 package io.github.lokka30.levelledmobs;
 
+import io.github.lokka30.levelledmobs.utils.Addition;
 import org.bukkit.Material;
 import org.bukkit.attribute.Attribute;
 import org.bukkit.entity.EntityType;
@@ -25,16 +26,7 @@ public class MobDataManager {
         }
     }
 
-    public void applyAttributeNewValue(LivingEntity entity, Attribute attribute, Double newValue) {
-        Objects.requireNonNull(entity.getAttribute(attribute)).setBaseValue(newValue);
-    }
-
-    public void setAttributeAddedValue(LivingEntity entity, Attribute attribute, Double addition, int level) {
-        double defaultValue = (double) getAttributeDefaultValue(entity.getType(), attribute);
-        applyAttributeNewValue(entity, attribute, defaultValue + (addition * level));
-    }
-
-    public boolean isLevelledDropManaged(EntityType entityType, Material material) {
+    public final boolean isLevelledDropManaged(EntityType entityType, Material material) {
         // Head drops
         if (material.toString().endsWith("_HEAD") || material.toString().endsWith("_SKULL")) {
             if (!instance.settingsCfg.getBoolean("mobs-multiply-head-drops")) {
@@ -44,5 +36,19 @@ public class MobDataManager {
 
         // Check list
         return instance.dropsCfg.getStringList(entityType.toString()).contains(material.toString());
+    }
+
+    public void setAdditionsForLevel(LivingEntity livingEntity, Attribute attribute, Addition addition, int currentLevel) {
+        final double defaultValue = (double) getAttributeDefaultValue(livingEntity.getType(), attribute);
+        Objects.requireNonNull(livingEntity.getAttribute(attribute)).setBaseValue(defaultValue + getAdditionsForLevel(livingEntity, addition, currentLevel));
+    }
+
+    public final double getAdditionsForLevel(LivingEntity livingEntity, Addition addition, int currentLevel) {
+        final int minLevel = instance.settingsCfg.getInt("fine-tuning.min-level");
+        final int maxLevel = instance.settingsCfg.getInt("fine-tuning.max-level");
+        final double range = (double) maxLevel - (minLevel - 1);
+        final double percent = currentLevel / range;
+        final double max = instance.settingsCfg.getDouble(addition.getMaxAdditionConfigPath());
+        return percent * max;
     }
 }